Feature/28158 pipeline refactor logic (#1469)
#28158 - ginormous refactor to help move us away from a coordinator-centric model

+++ polaris-gateway/TelemetryEvents/DocumentReclassifiedEvent.cs
@@ -3,7 +3,7 @@ namespace PolarisGateway.TelemetryEvents
     public class DocumentReclassifiedEvent : BaseRequestEvent
     {
         public DocumentReclassifiedEvent(
-            long caseId,
+            int caseId,
             string documentId)
         {
             CaseId = caseId;
